Refatora fluxo de login
